Add files via upload
integrate - 
I employed three methods—Midpoint, Trapezoidal Rule, and Simpson's Rule—to tackle this task. For n=1, the Midpoint method is used, approximating the area based on the middle point of the range. Trapezoidal Rule is applied for n=2, offering increased accuracy by incorporating two points. For n > 2, Simpson's Rule is the most accurate, applied with -1n points if n is odd, and -2n points to avoid exceeding the maximum sampling limit for even n.

areabetween- 
The method detects intersection points of functions 1f and 2f using the bisection method. The chosen range of 1 to 100 aims to strike a balance between identifying sufficient roots and computational efficiency. By calculating f = abs(f1 - f2), we measure the magnitude of their difference. After sorting intersection points, the integrate function is employed iteratively, using consecutive pairs of intersection points as ranges. The final result, representing the area, is an accumulation of the absolute values of the returned values. A value of 100 is selected for n in the integrate function to ensure both accuracy and reasonable runtime for the Simpson's Rule algorithm.