Add files via upload
area - 
In this assignment, I applied the Shoelace Formula method. Initially, I sampled 1000 points along the shape's outline. After confirming a minimum of three sampled points, the formula is used to calculate the area accurately. The choice of 1000 points was determined through testing to minimize errors in the area calculation. This mathematical algorithm calculates the area of a simple polygon based on the coordinates of its vertices.

fit_shape-
In this assignment, a maximum of 10000 points are initially sampled, with the aim of adhering to the 5-second runtime constraint. The sampling process is structured to ensure both a substantial number of points and adequate time for continued execution. Should the time limit result in fewer than 3 points being sampled, a Shape My is generated, and since an area cannot be accurately calculated with less than 3 points, a zero area is returned. For instances where sufficient points are sampled, the code continues with its usual procedures, including creating clusters, determining their centers, and arranging them in a coherent order. The final outcome involves constructing an object of type MyShape that receives the sorted clusters. This MyShape class inherits from the abstract class AbstractShape and effectively represents the shape. The class encompasses methods for contour representation and area calculation using the Shoelace Formula.